first attempt with interfaces and reflect everywhere
