+++ example/generic_go_function/makefile
@@ -1,0 +1,6 @@
+build: main
+
+main: main.go
+	go build -o main main.go
+
+.PHONY: all build main
